Assert with integer status codes to avoid hidden errors
We are using `resp['status']` which is a string ('202').
Instead, use `resp.status` which is an int. This is a problem
because `self.expected_success(202, '201')` doesn't raise an
error.

This also adds an `expected_success` method to the proper base
class to intercept calls to this method. This lets us warn about
this and cast to int.

And:
- check for 202 (not 200) on a zone update.
- add a .testr.conf so `ostestr` runs all the tests in the
current dir

Change-Id: I102c35eb6da0c5f99fd548b169ce490027fdb981