Add a zone import scenario test
This adds a test that imports a zone and checks the zone is created and
goes to active.

The base client's create/update calls now accept either a dict or string
as request data. Serialization is skipped if the data is a string. This
fixes a bug in the tests causing all zone import's to go to error.

Change-Id: I71d406efc5f2c9c35eae50761446c00a328ec5a2